requirements.txt
# Electricity Price Predictor - Dependencies
#
# Python Version: 3.9+ (tested on 3.9, 3.10, 3.11)
# Installation: pip install -r requirements.txt
#
# Note: You may see a urllib3 version warning - this is non-breaking and can be ignored.

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=10.0.0  # For Parquet files
requests-cache
retry-requests

# Machine learning
scikit-learn>=1.4.0
xgboost>=2.0.0
tensorflow>=2.15.0  # For LSTM model

# Data collection
openmeteo-requests>=1.0.0
requests>=2.31.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# UI (optional)
gradio>=4.0.0

# Production mode (Hopsworks Feature Store)
# Using unpinned version (installs latest compatible: 4.6.0+)
hopsworks==4.2.*
hopsworks[python]

# Pin boto3/botocore to prevent pip backtracking on urllib3 conflict
# (Non-breaking: works with urllib3 2.x despite conservative constraint)
boto3==1.42.7
botocore==1.42.7

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Development tools (optional)
# pytest>=7.4.0
# jupyter>=1.0.0
